Remote more dependencies from tests on ../extensions/
