+++ src/extension/commands/debug.ts
@@ -1,10 +1,11 @@
 import * as fs from "fs";
 import * as path from "path";
 import * as vs from "vscode";
+import { FlutterServiceExtension } from "../../shared/enums";
 import { Context } from "../../shared/workspace";
 import { Analytics } from "../analytics";
 import { CoverageData, PromiseCompleter } from "../debug/utils";
-import { FlutterServiceExtension, FlutterServiceExtensionArgs, FlutterVmServiceExtensions, timeDilationNormal, timeDilationSlow } from "../flutter/vm_service_extensions";
+import { FlutterServiceExtensionArgs, FlutterVmServiceExtensions, timeDilationNormal, timeDilationSlow } from "../flutter/vm_service_extensions";
 import { DebuggerType } from "../providers/debug_config_provider";
 import { PubGlobal } from "../pub/global";
 import { DevToolsManager } from "../sdk/dev_tools";
