+++ src/extension/flutter/hot_reload_save_handler.ts
@@ -1,10 +1,10 @@
 import * as path from "path";
 import { commands, debug, DiagnosticCollection, DiagnosticSeverity, ExtensionContext, workspace } from "vscode";
+import { restartReasonSave } from "../../shared/constants";
+import { FlutterService } from "../../shared/enums";
 import { DebugCommands } from "../commands/debug";
 import { config } from "../config";
-import { restartReasonSave } from "../../shared/constants";
 import { fsPath, isAnalyzableAndInWorkspace } from "../utils";
-import { FlutterService } from "./vm_service_extensions";
 
 export function setUpHotReloadOnSave(context: ExtensionContext, diagnostics: DiagnosticCollection, debugCommands: DebugCommands) {
 	let hotReloadDelayTimer: NodeJS.Timer | undefined;
