+++ src/extension/sdk/dev_tools.ts
@@ -3,12 +3,12 @@ import * as os from "os";
 import * as path from "path";
 import * as vs from "vscode";
 import { window, workspace } from "vscode";
+import { CHROME_OS_DEVTOOLS_PORT, pleaseReportBug } from "../../shared/constants";
+import { FlutterService } from "../../shared/enums";
 import { Analytics } from "../analytics";
 import { DebugCommands, debugSessions } from "../commands/debug";
 import { config } from "../config";
-import { CHROME_OS_DEVTOOLS_PORT, pleaseReportBug } from "../../shared/constants";
 import { isChromeOS, LogCategory } from "../debug/utils";
-import { FlutterService } from "../flutter/vm_service_extensions";
 import { PubGlobal } from "../pub/global";
 import { StdIOService, UnknownNotification } from "../services/stdio_service";
 import { getRandomInt, Sdks } from "../utils";
