+++ src/shared/enums.ts
@@ -10,3 +10,24 @@ export enum TestStatus {
 	Errored,
 	Running,
 }
+
+/// The service extensions we know about.
+export enum FlutterServiceExtension {
+	PlatformOverride = "ext.flutter.platformOverride",
+	DebugBanner = "ext.flutter.debugAllowBanner",
+	CheckElevations = "ext.flutter.debugCheckElevationsEnabled",
+	DebugPaint = "ext.flutter.debugPaint",
+	PaintBaselines = "ext.flutter.debugPaintBaselinesEnabled",
+	InspectorSelectMode = "ext.flutter.inspector.show",
+	InspectorSetPubRootDirectories = "ext.flutter.inspector.setPubRootDirectories",
+	RepaintRainbow = "ext.flutter.repaintRainbow",
+	PerformanceOverlay = "ext.flutter.showPerformanceOverlay",
+	SlowAnimations = "ext.flutter.timeDilation",
+}
+
+/// The service extensions we know about and allow toggling via commands.
+export enum FlutterService {
+	HotReload = "reloadSources",
+	HotRestart = "hotRestart",
+	LaunchDevTools = "launchDevTools",
+}
