+++ src/shared/interfaces.ts
@@ -1,11 +1,21 @@
 import { DebugSession, DebugSessionCustomEvent, TreeDataProvider, TreeItem } from "vscode";
-import { TestStatus } from "./enums";
+import { FlutterService, FlutterServiceExtension, TestStatus } from "./enums";
 
 export interface ITestResultsProvider extends TreeDataProvider<ITestItemTreeItem> {
-	handleDebugSessionCustomEvent(e: DebugSessionCustomEvent): void;
 	handleDebugSessionEnd(session: DebugSession): void;
+	handleDebugSessionCustomEvent(e: DebugSessionCustomEvent): void;
 }
 
 export interface ITestItemTreeItem extends TreeItem {
 	status: TestStatus;
+}
+
+export interface DebugCommandHandler {
+	flutterExtensions: {
+		serviceIsRegistered(service: FlutterService): boolean;
+		serviceExtensionIsLoaded(extension: FlutterServiceExtension): boolean;
+	};
+	handleDebugSessionStart(session: DebugSession): void;
+	handleDebugSessionEnd(session: DebugSession): void;
+	handleDebugSessionCustomEvent(e: DebugSessionCustomEvent): void;
 }
